Frequency Domain processing pull over
Trying to reimplement old frequency domain stuff at least for
pitch detection, could also be continued for shifting maybe.
This is good practice for somewhat cleanly implementing more
processor elements, especially if the detection works somewhat
as expected.

Final detection might still be through autocorrelation or
something in the end, but this compiles and I believe I can
get it to work for now.